refactor(ios): extract UI-test harness and chat placeholder from RootView (#14239)
## Summary

Behavior-preserving structural split of `RootView.swift` (**1,239 → 918 lines**), iteration 1 of the iOS refactor loop. No Linear issue — ad-hoc loop work.

- Move the DEBUG-only UI-test harness (`LiveAuthUITestStatus`, `LiveAuthCallbackLaunchInput`, `UITestExitButton`, `UITestingAuthCallbackRoot`) to a new `App/UITestHarnessViews.swift` so production view code no longer carries test scaffolding. All types remain `#if DEBUG`; access widened from `private` to internal only where cross-file references require it.
- Move `MobileChatPlaceholderView` + `ChatComposerPreview` to `Features/Chat/MobileChatPlaceholderView.swift` (chat placeholder now lives with the chat feature).
- Delete dead private `waitForClerkLoaded` helper (zero call sites).
- Wire both new files into `project.pbxproj` (explicit `PBXFileReference`/`PBXBuildFile` entries; validated with `plutil -lint`).

No logic changes; every moved declaration is verbatim.

## Verification

- `pnpm run ios:lint` → **clean ✓**
- `pnpm run ios:test` (iPhone 17 sim, iOS 26.3.1) → **248 passed, 6 skipped, 2 failed**
- The 2 failures (`testNoGhostFootprintAfterPillRemovalAcrossStates`, `testChatComposerPreservesDraftAcrossShellNavigation`) are **pre-existing on clean `main` (62d314ace)** — verified by stashing this diff and re-running: identical deterministic failures. Tracked in **JOV-4252**.
- Guard tests (`AuthScreenStyleGuardTests`, `SettingsEscapeTrapGuardTests`, `AppStateTests`, `MobileChatContentParserTests`) all pass.

## Risks

- Low: pure code move. The moved harness types are only referenced from DEBUG UI-testing launch modes (`JovieApp.swift` keeps its `#if DEBUG` gate).
